+++ noodles-vcf/src/header/parser/record/value.rs
@@ -71,7 +71,6 @@ pub(super) fn parse_value(src: &mut &[u8], key: Key) -> Result<Record, ParseErro
         key::CONTIG => map::parse_contig(src)
             .map(|(id, map)| Record::Contig(id, map))
             .map_err(ParseError::InvalidContig),
-        key::META => todo!(),
         Key::Other(k) => {
             let is_map = src.first().map(|&b| b == b'<').unwrap_or_default();